requirements.txt
absl-py==0.7.0
astor==0.7.1
bleach==1.5.0
certifi==2019.3.9
cycler==0.10.0
Cython==0.29.6
gast==0.2.2
grpcio==1.16.1
h5py==2.8.0
html5lib==0.9999999
Keras==2.1.5
keras-vggface==0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.3
mkl-fft==1.0.6
mkl-random==1.0.1
numpy==1.15.4
pandas==0.24.2
Pillow==5.4.1
protobuf==3.6.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
scipy==1.1.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
tornado==6.0.1
Werkzeug==0.14.1
